Move addToCollection modals to SearchResult component
