+++ src/app/components/SearchResult/SearchResult.tsx
@@ -1,9 +1,15 @@
 import { useScrollPosition } from '@n8tb1t/use-scroll-position';
+import axios from 'axios';
 import type { MutableRefObject } from 'react';
 import React, { useRef, useState } from 'react';
 import { isMobileOnly } from 'react-device-detect';
 import Masonry from 'react-masonry-component';
+import { useHistory } from 'react-router';
 import type { image, imagesResult } from '../../../lib/types/image';
+import useCurrentUser from '../../hooks/useCurrentUser';
+import Button from '../Button/Button';
+import Headline from '../Headline/Headline';
+import Icon from '../Icon/Icon';
 import ImageCollections from '../ImageCollections/ImageCollections';
 import ImageDetails from '../ImageDetails/ImageDetails';
 import Modal from '../Modal/Modal';
@@ -18,7 +24,6 @@ export type SearchResultProps = {
   isLoading: boolean;
   isFetchingNewResult?: boolean;
   imagesResult: imagesResult | null;
-  onCollectionClick: (id: number) => void;
   handleScroll: (position: number, parentHeight: number) => void;
   className?: string;
 };
@@ -53,11 +58,29 @@ const modalDetailsSize = {
   },
 };
 
+const popupSize = {
+  desktop: {
+    minHeight: '',
+    minWidth: '',
+    height: '',
+    width: '',
+    maxHeight: '',
+    maxWidth: '50%',
+  },
+  mobile: {
+    minHeight: '',
+    minWidth: '',
+    height: '',
+    width: '',
+    maxHeight: '',
+    maxWidth: '85%',
+  },
+};
+
 export default function SearchResult({
   isLoading,
   isFetchingNewResult = false,
   imagesResult,
-  onCollectionClick,
   handleScroll,
   className = '',
 }: SearchResultProps): JSX.Element {
@@ -67,6 +90,22 @@ export default function SearchResult({
   const [masonryComplete, setMasonryComplete] = useState<boolean>(false);
   const searchResultElement = useRef<HTMLElement>(null);
   const searchResultEndElement = useRef<HTMLDivElement>(null);
+
+  const [addToCollectionResult, setAddToCollectionResult] = useState<{
+    success: boolean;
+    message: JSX.Element;
+    collectionName?: string;
+    collectionId?: string;
+    image?: image;
+  } | null>(null);
+  const [modalAddToCollectionExtended, setModalAddToCollectionExtended] = useState<boolean | null>(
+    null
+  );
+  const modalAddToCollectionExtendedRef = useRef(modalAddToCollectionExtended);
+  modalAddToCollectionExtendedRef.current = modalAddToCollectionExtended;
+
+  const currentUser = useCurrentUser();
+  const history = useHistory();
 
   useScrollPosition(
     ({ currPos }) => handleScroll(currPos.y, searchResultElement.current?.offsetHeight || 0),
@@ -95,6 +134,52 @@ export default function SearchResult({
     );
   }
 
+  async function handleAddToCollection(image: image | undefined): Promise<void> {
+    if (!image) return;
+    if (!currentUser) history.push(`/profile`);
+
+    try {
+      const addToCollectionResult = await axios.post('/api/images', {
+        image,
+      });
+      const { collectionId, collectionName } = addToCollectionResult.data?.result;
+      setAddToCollectionResult({
+        success: true,
+        message: (
+          <>
+            Added to collection <strong>{collectionName}</strong>
+          </>
+        ),
+        collectionId,
+        collectionName,
+        image,
+      });
+
+      setModalAddToCollectionExtended(false);
+      setTimeout(() => {
+        if (modalAddToCollectionExtendedRef.current === false)
+          setModalAddToCollectionExtended(null);
+      }, 4000);
+    } catch (error) {
+      if (axios.isAxiosError(error)) {
+        if (error.response?.data?.resultCode === 501) {
+          setAddToCollectionResult({
+            success: false,
+            message: <>No collection found!</>,
+            image,
+          });
+        } else {
+          setAddToCollectionResult({
+            success: false,
+            message: <>Failed to add image!</>,
+            image,
+          });
+        }
+      }
+      setModalAddToCollectionExtended(true);
+    }
+  }
+
   return (
     <section className={`${styles.imagesList} ${className}`} ref={searchResultElement}>
       <Spinner show={!masonryComplete} className={styles.spinner} />
@@ -119,7 +204,7 @@ export default function SearchResult({
                 setSelectedImage(imagesResult?.results[index] || null);
                 setModalActiveTab('details');
               }}
-              onCollectionClick={() => onCollectionClick(index)}
+              onCollectionClick={() => handleAddToCollection(image)}
               className={styles.searchResultImage}
               key={image.id}
             />
@@ -147,6 +232,42 @@ export default function SearchResult({
             {modalActiveTab === 'download' && 'Download'}
             {modalActiveTab === 'palette' && 'palette'}
           </div>
+        </div>
+      </Modal>
+
+      <Modal
+        show={modalAddToCollectionExtended === false}
+        position="bottomRightSlide"
+        size={popupSize}
+        backgroundOverlay={false}
+      >
+        <div className={styles.popupSmall}>
+          <Icon icon={addToCollectionResult?.success ? 'check' : 'close'} color="mediumGradient" />
+          <div>{addToCollectionResult?.message}</div>
+          {addToCollectionResult?.success && (
+            <Button
+              icon="edit"
+              text="Change"
+              small
+              onClick={() => setModalAddToCollectionExtended(true)}
+            />
+          )}
+        </div>
+      </Modal>
+
+      <Modal
+        show={modalAddToCollectionExtended === true}
+        position="bottomRight"
+        size={popupSize}
+        backgroundOverlay={false}
+      >
+        <div className={styles.popupLarge}>
+          <Headline className={styles.headline}>Add to</Headline>
+          <ImageCollections
+            className={styles.collections}
+            image={addToCollectionResult?.image || null}
+          />
+          <Button icon="check" text="Done" onClick={() => setModalAddToCollectionExtended(null)} />
         </div>
       </Modal>
     </section>
