+++ src/app/pages/Search/Search.module.css
@@ -8,38 +8,3 @@
 .input {
   margin: var(--gapSmall);
 }
-
-.popupSmall {
-  display: grid;
-  grid: 100% / auto 1fr auto;
-  margin: 1rem;
-  gap: 1rem;
-}
-
-.popupSmall * {
-  align-self: center;
-}
-
-.popupLarge {
-  display: grid;
-  margin: var(--gap);
-  gap: var(--gap);
-  justify-items: self-end;
-}
-
-.headline {
-  width: 100%;
-}
-
-.collections {
-  max-height: 70vh;
-  overflow-y: auto;
-}
-
-.collections::-webkit-scrollbar {
-  width: var(--gapSmall);
-}
-.collections::-webkit-scrollbar-thumb {
-  background: var(--primary);
-  border-radius: var(--radius);
-}
